+++ lib/main.dart
@@ -8,7 +8,7 @@ import 'package:solarisdemo/infrastructure/transactions/transaction_service.dart
 import 'package:solarisdemo/ivory_app.dart';
 import 'package:solarisdemo/redux/app_state.dart';
 import 'package:solarisdemo/redux/store_factory.dart';
-import 'package:solarisdemo/services/push_notifications_service.dart';
+import 'package:solarisdemo/services/push_notification_service.dart';
 
 import '../config.dart';
 import 'firebase_options.dart';
