+++ lib/redux/store_factory.dart
@@ -3,7 +3,7 @@ import 'package:solarisdemo/infrastructure/notifications/notifications_middlewar
 import 'package:solarisdemo/infrastructure/transactions/transaction_middleware.dart';
 import 'package:solarisdemo/infrastructure/transactions/transaction_service.dart';
 import 'package:solarisdemo/redux/app_reducer.dart';
-import 'package:solarisdemo/services/push_notifications_service.dart';
+import 'package:solarisdemo/services/push_notification_service.dart';
 
 import 'app_state.dart';
